+++ web/app/components/core/not-found.component.jsx
@@ -5,7 +5,7 @@ var notFound = {};
 notFound.displayName = 'NotFound';
 
 notFound.render = function notFoundComponentRender() {
-  return <h2>Not Found</h2>;
+  return <h1>Not Found</h1>;
 };
 
 module.exports = React.createClass(notFound);