+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # master
 
+- refactored mocking solution on front-end to be a little more flexible
+- added nock for mocking within unit tests that uses the same data as the front-end mocks
+- fixed a few with-reolves unit tests
+- mocked requests should now properly support request headers
+- added ability to find rendered dom component by prop value
+- updated mocking library to support request header checking properly
+- added jsuri library
 - updated gulp clean build task to remove coverage generated files
 - better support for testing react router based components
 - assets rewrite add `/build/` to path with configurable option to add `/static/[SHA]/`